+++ ogma-language-jsonspec/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for ogma-language-jsonspec
 
+## [1.3.0] - 2024-03-21
+
+* Version bump 1.3.0 (#133).
+* Make some fields in JSON format spec optional (#122).
+
 ## [1.2.0] - 2024-01-21
 
 * Version bump 1.2.0 (#117).